+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Improved error message when no concrete binded class was found
+
 ## 0.5.0
 ### 2023-05-19